python-services/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.0
pandas==1.5.0
scikit-learn==1.3.0
mysql-connector-python==8.2.0
pymysql==1.1.0
sqlalchemy==2.0.23
pillow==10.0.0
scipy==1.11.4
python-multipart==0.0.6